In constructor, Decoder now grabs mod nonterminal maps from its DecoderServer object.  The use of these maps prevent the computation of mod nonterminal and mod nonterminal head-word probabilities when the current future for the mod nonterminal model is impossible given the last-level back-off history.
The constructor has been (temporariliy) modified to
output the prunedPretermsPosMap to System.err.

The data member pruneFact is now computed differently,
since the user is now to specify it in base 10 (instead of e).

Added an egregious HACK to *not* prune away apostrophes
(i.e., words whose string representation consists solely
of a single apostrophe character), since this was messing
up a lot of sentences.

The complete method has been converted back to a
pair of double loops.