In collectStats method, changed calls to Subcat.getCanonical to reflect new method signature.
In collectModifierStats method, removed unnecessary copying of dynamic subcat;
copying now only happens when the subcat needs to be modified (because it
contains the current modifier nonterminal label or gap requirement).

In deriveCounts method, changed name of local canonicalEventLists variable to
be simply canonical.  Also, removed all calls to Model.canonicalize, as
canonicalization now happens within the call to Model.deriveCounts, the calls
to which were updated to reflect the new signature, where the third argument is
the canonicalization map.

Modified deriveSubcatMaps method to reflect new signature of
Subcat.getCanonical method.