* Fixed MAJOR bug in the containsVerb() method, whereby if head child contained a verb in its surface string but it was *not* the head word of the head child, the predicate would return a false negative.  Because of this, the containsVerb predicate didn't properly handle a crucial part of the recursive definition of containing verbs in the surface string.  As part of the fix, every chart item now has a new containsVerb data member (of type byte), whose value is one of three new constants: containsVerbUndefined, containsVerbTrue and containsVerbFalse.
* Modified headLabel() method of CKYItem$BaseNPAware class, to duplicate Mike
Collins' bizarre behavior whereby whenever the head child of an item is a base
NP, the head label of that item is taken to be the label of the head child of
the NPB.  For example, if you have a chart item representing the subtree
	(NP (NPB (NNP John)))
the head label of NP is taken to be NNP, and not NPB.  While we duplicate
Collins' odd behavior here, it is *not* duplicated in the much more frequently
used class CKYItem$MappedPrevModBaseNPAware.

* Now uses new headLabel() method in various equality methods, which simplifies
code (and makes it a tad more readable).

* Added Constraint field to chart items, along with accessor/mutator methods
for that field.