* Added new debugging options: debugPrunedPretermsPosMap and debugPrunedPunctuationPosMap.
* Added debugOutputAllCounts debugging option.

* Added various Symbol constants for use when debugging.

* Added data member posSet, which is a Set object containing all possible part
of speech Symbol objects, derived from the values contained in the part of
speech map (which maps words to their possible parts of speech, and is obtained
by invoking DecoderServerRemote.posMap).

* Added new data members to cache various static symbols (such as
the value returned by Treebank.baseNPLabel()).

* Added two new arrays of booleans, commaForPruning and conjForPruning, which
are set via the setCommaConstraintData() method (invoked within the
initialize(SexpList,SexpList) method).  These arrays cache the locations of
commas and conjunctions that are to be used when determining whether a chart
item violates the comma pruning constraint.  Accordingly, these arrays are
(currently) used only by the commaConstraintViolation(int,int,int) method.

* Added call to static method Training.setUpFastArgMap, so that argument
identification is simply a hash-lookup away.

* Modified postProcess(Sexp) method so that it only invokes the private
restoreOriginalWords method and then, if debugDontPostProcess is false, acts as
a flow-through invocation of the new Training.postProcess method.

* Modified initialize method so that it spits out a warning message if a part
of speech tag supplied for a word was never seen during training.  Also, there
is a new setting useLowFreqTags that can be set during training and decoding.
If it is, this method will use the tags observed for unknown words seen during
training along with any supplied tags as the set of possible tags for these
words.  E.g., if the word "foobar" was seen in training with the tags NNP and
DT, and was supplied in a sentence with the tag JJ (as in (foobar (JJ))), then
the chart will be seeded with items whose headwords are foobar/NNP and
foobar/DT, as well as foobar/JJ.  If the value of Settings.useLowFreqTags is
false, then only the supplied tag (JJ, in the example) will be used.

* Added setUnion utility method for use by the new useLowFreqs behavior
implemented in the initialize method (described above).

* Added (as yet unused) ConstraintSet argument to parse method.

* Made small changes to the debugAnalyzeChart block of the parse method.

* Changed complete(int,int) so that even if there is a comma constraint
violation, we still try to find modificands that have not yet received their
stop probabilities whose labels are ANPB, to see if we can add a premodifier
(so that we can build base NPs to the left even if they contain commas.

* Fixed SERIOUS BUG in joinItems method, involving off-by-one error in the way
previous mods were being computed.

* Fixed SERIUOS BUG in addStopProbs method, involving off-by-one error in the
way previous mods were being computed.

* Added calls to new static Shifter.skip(Item,Sexp,Sexp) method in
getPrevMods method, so that various nonterminals may be skipped, if desired,
when constructing a previous-modifier list.

* Added calls to new static Shifter.skip(Item,Word,Word) method in getPrevModWords method, so that various modifier head words may be skipped, if desired,
when constructing a previous-mod-head-word list.

* Changed commaConstraintViolation to use new commaForPruning and
conjForPruning arrays (described above).

* Various other, small changes.