* Not truly necessary, but added fall-back default for head table resource, in case Settings.headTableResource is not set (should never be the case, since default-settings.properties shoudl ALWAYS be supplied with the parser's jar).
* The method defaultFindHead(Symbol,SexpList) now spits out a warning
when a rule cannot be found and when there is no default rule.  Before, it
used to assume that a default rule existed.

* Modified helper method tagMatches(Symbol,Symbol[]) so that it now uses the
Nonterminal class to determine whether the tag supplied in a head-rule
(confusingly called a "matchTag" in the code) subsumes the tag we're trying to
match in the right-hand side of a production.  This way, e.g., when a user
writes the rule (NP (r NP)) it will match an NP-A, an NP-SBJ and all other
NPs that might occur on the rhs of a production.