* Overrode lambdaFudge(int) so that it returns 0.0 for all back-off levels.
* Overrode lambdaFudgeTerm(int) so that it returns 1.0 for all back-off levels.

* Crucially modified getFutrue(TrainerEvent event, int backOffLevel) method so
that it checks the PriorEvent's head word (a Word object) to see if its
features field is non-null.  If so, it uses that as the value of the "word";
otherwise, it uses the word field.  This enables the transparent handling of a
"keep all words" feature as Collins implements it.