* Added the necessary side element to all histories constructed by getHistory method (see comment below regarding the removeHistory method).
* Overrode doCleanup() method to return true.

* Overrode the default, do-nothing implementation of the
removeHistory(int,Event) method.  Here is the comment explaining why.
The trainer passes ModifierEvent objects representing modifiers on both sides
of the head, even though there are separate models for each side.  This allows
the final back-off level, p(w | t), to use information gathered from both
sides.  However, it means that after all counts have been derived, there are
twice as many entries in the counts tables for the back-off levels 0 and 1,
since the model for, say, the left side will have stored counts for the right
side as well (the counts are stored separately by explicitly including a "side"
field in the events being counted).  This method allows for the removal of
these "unnecessary" counts, which will never be used when decoding.

* Updated copy method to include crucial additionalData field, which indicates
which side of the head this probability structure is modeling.