* Now performs deficient estimation as is done in the Collins parser based on value of the Settings.collinsDeficientEstimation property.
* Now uses a lambda fudge term when smoothing probabilities, as is done in the
Collins parser.  Formula for computing lambdas is
	      s
	-------------
	s + C1*u + C2
where s is the number of occurnces of the history, u is the number of unique
outcomes observed and C1 is the labmda fudge factor and C2 is the lambda fudge
term.

* Added new cleanup method.  See javadoc comment for
ModWordModelStructure2.removeHistory(int,Event) to explain how and why this is
used.

* Better debugging output when debugging level is >= 21.

* Implemented share method, which allows counts from one Model to be shared
with another, thereby allowing the sort of hack that Collins did when he shared
all counts that estimated p(w | t) with all models that needed to estimate that
probability.  (Of course, his mistake was not to take into account the fact
that some part of speech tags "t" are part of lexicalized nonterminals, whereas
others are simply the preterminals.)