* Removed all traces of anything having to do with WordNet, as those data and methods will (and should) all be in a subclass.  The only traces left, however, were those appearing in out-dated javadoc comments.
* Added new features data member of type Symbol, which is used for, among other
things, producing the desired behavior when keepAllWords is true.

* Added accessor and mutator methods for new features data member.

* Added several javadoc comments that were missing.

* Modified constructor Word(Sexp) so that not only accepts lists with two
symbols, but also those with three symbols, where the third symbol is the value
for the Word object's features field.

* Modified equals and copy methods to take into account new features data
member.  Note that the hash code method was *not* modified, as the hash codes
produces will still be consistent with equals (but not modifying hashCode may
yield larger hash buckets).

* Adjusted toSymbol to output features field when it is non-null.