* Updated class-level javadoc comment to more accurately reflect functionality of this class.
* Added int[] collinsSkipArr constant data member, which contains hard-coded
numbers of sentences that Collins' trainer skips due to their having more than
500 tokens, where a token is a word, a nonterminal or a parenthesis (a behavior
introduced by one of Collins' preprocessing Perl scripts, where the comment
reads something like "this is done for strange, historical reasons").  This
behavior is emulated when the value of Settings.collinsSkipWSJSentences is
true.

* The text file that the main method produces containing all top-level
observations and their counts now contains two new items: pruned-preterms and
pruned-punc.  The former is the set of all pruned preterminals, and the latter
is the set of all punctuation preterminals that were "raised away" by
Training.raisePunctuation(Sexp) because they occurred either at the beginning
or at the end of the sentence.

* Modified alterLowFrequencyWords so that when Settings.keepLowFreqTags is
true, it will add low frequency words to the pos map it creates.  This enables
the Decoder to use that information when decoding (if Settings.useLowFreqTags
is true).

* The train(SexpTokenizer,boolean,boolean) method now invokes Training.skip on
each input sentence to determine whether it should be skipped or not.  This
skipping is separate from the skipping performed when
Settings.collinsSkipWSJSentences is true.

* Modified collectStats(Sexp,HeadTreeNode,boolean) so that when it is
collecting stats on the root of the tree, it manufactures a ModifierEvent
object as though the root node of the observed tree "modifies" the magical
+TOP+ hyper-root node, so as to gather counts for the last back-off level of
the modificatio model, which is p(w | t).  In other words, this enables that
back-off level to obtain counts for words that are the heads of entire
sentences, which are normally not modifiying anything, and are therefore not
ever counted as futures in the word-bigram models p(w_m | ..., w_h, ...).

* Implemented behavior to emulate Collins' trainer, which is activated when
outputCollins constant is set to true.

* Modified collectModifierStats(HeadTreeNode,Subcat,int,boolean) so that it
uses Shifter to shift previous modifier nonterminals and previous modifier head
words into SexpList and WordList state objects, respectively.

* Modified deriveCounts() method to invoke the setAdditionalData method for the
ProbabilityStructure objects contained within the Model objects for the left-
and right-modifier and left- and right-modifier word generation models.  This
makes these ProbabilityStructure objects "aware" of which side of the head they
model generation.  See javadoc comment for
ModWordModelStructure2.removeHistory(int,Event) method to find out why this is
useful and necessary.

* Hard-coded count-sharing to emulate Collins' behavior, where the same counts
are used whenever any parameter class needs to compute p(w | t), even if what
is really needed is the computation of p(w | t_NT), where t_NT is in the space
of tags appearing in lexicalized nonterminals.  We need to devise a reasonable
method to specify such count-sharing, perhaps as run-time code read in (via a
bsh.Interpreter instance?) by the ModelCollection class.