* Made sweeping change: the ModelCollection object held by an instance of the DecoderServer class now uses only a single Model object for modifying nonterminals and a single Model object for head words of (lexicalized) modifying nonterminals; before, two Model objects were used for each, one for modifiers on the left and one for modifiers on the right side of the head. Accordingly, we have un-commented the cases in the getHistory and getBaseNPHistory methods which previously added the side component, so that these models can distinguish, where necessary, modifiers on the left and right sides.  Crucially, the p(w | t) back-off level has no side component, meaning that modifier words appearing on either side of the head word are used for estimating the probabilities at that level.
* The Trainer class now adds a "fake" modifier event, where the root node of an
observed tree "modifies" +TOP+, so that the p(w | t) estimates can include
counts for words that are the heads of entire sentences, and are therefore
themselves not modifiers of any head; this fake event forces them to be the
modifier word of something.  While this has the desired effect of making the
p(w | t) back-off level use all possible counts, it has the undesirable effect
of causing the Model class to derive useless counts at the more-detailed
back-off levels from this "fake" modifier event.  In order to counteract this
undesirable consequence, we made the following two changes to this class:
	* we overrode the doCleanup event to return true
	* we overrode the removeHistory event so that it returns true when it
	  encounters an event derived from the "fake" modifier event