* Previously, this class overrode the doCleanup method to return true and overrode the removeHistory method so that it could rid itself of counts that did not apply to the side-specific Model for which it was helping compute probabilities.  Since there are no longer side-specific models for modifier words, the removeHistory method has been altered.
* The Trainer class now adds a "fake" modifier event, where the root node of an
observed tree "modifies" +TOP+, so that the p(w | t) estimates can include
counts for words that are the heads of entire sentences, and are therefore
themselves not modifiers of any head; this fake event forces them to be the
modifier word of something.  While this has the desired effect of making the
p(w | t) back-off level use all possible counts, it has the undesirable effect
of causing the Model class to derive useless counts at the more-detailed
back-off levels from this "fake" modifier event.  In order to counteract this
undesirable consequence, we modified the removeHistory method so that it
returns true when it encounters an event derived from the "fake" modifier
event.