* Now that CountsTable objects record double's instead of int's, the return value of vocabCounter.count(word) (line 143) is cast to an int (since we never have fractional word counts).
* Removed leftModNonterminalMap and rightModNonterminalMap accessor methods,
replacing them with single modNonterminalMap accessor method, since we now
depend on direction, if desired as a dependency, to be present in events at
last level of back-off in mod nonterminal model.  The interface
DecoderServerRemote was changed accordingly.

* Added probHead, probMod, probLeftSubcat, probRightSubcat and probTop methods,
which are non-log prob versions of their log-prob analogs.  This allows the
EMDecoder class to use real probabilities, instead of log probs.