* Changed the type of the eventsToTypes data member from Map to MapToPrimitive, allowing the mapping of the types directly to int's (which is what we want; previously, we mapped the type objects to Integer objects, from which we had to un-wrap the int values when reading in an observations file).
* Replaced leftModNonterminalMap and rightModNonterminalMap data members with a
single new data member, modNonterminalMap, since we now use direction elements
in the event objects themselves if we wish direction to be a dependency (see
Decoder.futurePossible).  Accordingly, changed name of deriveModNonterminalMaps
to be deriveModNonterminalMap (since there is only one map now).  Finally, the
name of outputModNonterminalMaps was appropriately changed to be
outputModNonterminalMap.

* Since ModelColleciton.set method also replaced the two direction-sensitive
map arguments with a single argument, the invocation was appropriately changed.

* Changed readStats method so that it appropriate calls Double.parseDouble in
order to read counts, since counts are now allowed to be fractional (to allow,
among other things, proper EM training).  Previously, the count strings were
converted to numbers using Integer.parseInt.

* outputCollins method assumes counts are integers, and now calls
entry.getDoubleValue() (as it should), but casts the return value to an int.