* Constructor no longer calls Treebank.getCanonical on nonterminal labels, since it is now assumed that labels used to construct a tree of constraints have already been appropriately pre-processed.
* Fixed bug in isSatisfiedBy method, whereby satisfied data member could go
from being false to being true and then *back* to being false again, which
should never happen, since once a constraint is satisfied, it cannot be
"un"-satisfied.

* Changed isLocallySatisfiedBy method so that it first checks if the item's
label is object-equal to the constraint's label and *then* checks if the
canonical version of the item's label is object equal to the constraint's
label.  By doing this, new labels introduced via preprocessing of training
trees (such as SG in English) will be appropriately matched to their
non-transformed variants (such as S in the English Treebank).  This method of
determining label equality also guarantees that if the constraint *does*
contain a transformed label, such as NPB, then it will indeed be matched to the
item's label.  Ideally, though, we should create a LabelComparator interface,
allowing arbitrary specification of how to test for equality for such classes
as this one (i.e., you optionally pass in a LabelComparator upon construction
of an UnlexTreeConstraint object).