* Replaced the type of the countThreshold data member from int to double. Accordingly, changed decoding of Settings.countThreshold property to use Double.parseDouble instead of Integer.parseInt.
* Added derivedCountThreshold, whose value is specified via the
Settings.derivedCountThreshold property, decoded using Double.parseDouble.

* Changed train method so that, just before end of its execution, if
countThreshold is greater than 0.0, it removes all head events, modifier events
and gap events from the appropriate tables that are lower than the threshold.

* Changed deriveCounts method so that it passes a double corresponding to the
value of the Settings.derivedCountThreshold property to the new definition of
the MOdel.deriveCounts method, which takes a double as its third parameter
instead of an int as before.

* Changed readStats(SexpTokenizer) so that it checks count of all head,
modifier and gap events read in and discards the events if they are below the
count threshold, as specified via the Settings.countThreshold property.