* Added timeout feature, utilizing TimeoutException inner class.
* Added headToParentMap.

* Changed values of subcat maps to be arrays, for less expensive iteration
during parsing.

* Added option for local caching decoder server.

* Made word removal (in preProcess method) less error-prone, by factoring
out word-removal code into its own method, removeWord(SexpList,SexpList,int).

* Added support for language-specific sentence preprocessing, with a call to
Language.training.preProcessTest at the end of the preProcess method.

* Factored out the complicated code that determines what set of tags to use
when seeding the chart for a particular word.  The code is now in the new
method, getTagSet.  This method issues a warning to System.err if the
useOnlySuppliedTags setting is true but the word was never seen with any of the
supplied tags.

* Factored out code that actually adds items to the chart for a particular
word of the test sentence into a new method called seedChart.

* Modified addUnaries so that, when useHeadToParentMap is true, it checks
the headToParentMap and only builds with nonterminals that have been seen
as parents of the specified chart item's label.