* Added parserClass data member, so that subclasses can easily re-use code of main method, by first setting this data member to be the subclass, and then invoking main.
* Replaced "new DecoderServer" in constructor with call to new method
"getNewDecoderServer", which allows for subclasses to get an instance of
DecoderServerRemote of their choosing by overriding the getNewDecoderServer
method.

* In parse(SexpList sent) method, replaced call to ConstraintSets.get
with a call to the new method getConstraintsFromTree, allowing subclasses more
control over how constraints are extracted.

* Added convertUnknownWords method, which transforms all low-frequency words in
a tree to their appropriate "unknown" versions (whether +UNKNOWN+ or a word
feature vector).  This method is called by getConstraintsFromTree before
actually getting the constraints (using ConstraintSets.get), so that if the
constraints depend on lexical items, they will be the same lexical items that
the decoder would be dealing with.

* Changed several methods from private to protected access level:
	getRetries, getRetrySleep, getFailover, getIntProperty

* Made various changes to main method.