* Now uses Settings.trainerShareCounts to determine whether to call the new method ModelColleciton.shareCounts.
* Added static data member trainerClass, which allows subclasses to re-use
the main method, by first setting this data member to be the subclass
and then invoking Trainer.main from <subclass>.main.

* Changed access level of all instance data members from private to public.

* The internal ModelCollection is no longer created directly using the new
operator, but rather by invoking the new method newModelCollection, which
allows subclasses to sneakily hand back a subclass of ModelCollection when
the Trainer requests one.

* Added six helper methods for adding TrainerEvent objects to the headEvents,
modifierEvents and gapEvents data members.  These helper are called
addHeadEvent, addModifierEvent and addGapEvent (two versions each), and are
called by the collectStats and readStats methods.  These methods provide a hook
for subclasses.

* Changed access level of collectStats from private to protected.

* In collectStats method, changed calls to headEvents.add, modifierEvents.add
and gapEvents.add to be calls to the new helper methods addHeadEvent,
addModifierEvent and addGapEvent (see comment above).

* Changed access level of collectModifierStats from private to protected.

* In collectModifierStats method, changed calls to modifierEvents.add and
gapEvents.add to be calls to addModifierEvent and addGapEvent, respectively.

* Changed access level of addToPosMap from private to protected.

* In deriveCounts method, a new method is used to get Model objects.  Before, a
Model object was created directly using the new operator, and a
ProbabilityStructure was passed into the Model object's constructor as an
argument.  The new method calls ProbabilityStructure.newModel to a get a Model
object wrapped around the ProbabilityStructure object.  This allows
ProbabilityStructure objects to create the Model instance appropriate for them
(such as JointModel instead of Model).

* The deriveCounts method now calls the new method deriveHeadToParentMap.

* Added a new method deriveCountsHook, which is called at end of deriveCounts
method just prior to setting priorEvents, headEvents, modifierEvents and
gapEvents data members to null.  This is obviously for subclasses.

* Created helper method for deriveCounts (also named deriveCounts, via
overloading) that makes the actual calls to the respective Model objects'
deriveCounts methods.

* Added readStatsHook and writeStatsHook methods, for subclasses to write out
and read in extra data to the observed data file, if desired.

* Changed usgae of main method so that settings file is specified using
"standard" command-line switch -sf (or --settings).