* Fixed a bug in IOData constructor whereby if the specified output filename was null (legitimate if the user doesn't want to use Switchboard's object-processing facilities), it would still try to create a File object around the null filename.  A simlar bug was fixed w.r.t. the log filename.
* Added IOData.registerConsumer method (flow-through from new
Switchboard.registerConsumer method).

* IOData.writeToLog is now synchronized.  Also, this method now checks to make
sure that the log filename that was passed to IOData's constructor is non-null.
Finally, before actually writing to the log file, this method grabs a lock on
the list of consumers and makes sure they consume the current object before its
processed version, if any, is written to file.

* Changed KeepAlive.run method so that if a call to the switchboard user's
keepAlive method fails, it waits keepAliveInterval milliseconds before trying
again.

* Switchboard now has a List of Consumer objects.

* Modified Switchboard.export method so that it ignores custom RMI socket
factories (doesn't pass them on to UnicastRemoteObject.export method
invocation).  This is because TimeoutSocketFactory has a bug that was causing
recent (1.4.x) versions of RMI to hang.

* Fixed processFile so that a dumper thread is only started if the user has
specified a non-null log filename and output filename.

* The Switchboard.nextObject method now calls Consumer.processingComplete on
every consumer in its private list when the current file's processing is
complete.

* Added Switchboard.registerConsumer method.