* Added deleteCountsWhenPrecomputingProbs boolean constant, to control whether counts data member is nullified before saving Model to file (or otherwise serialized).  It can be useful *not* to nullify counts (keep raw counts around) even when precomputing probabilites for debugging and analysis purposes (see AnalyzeDisns class).
* Added saveBackOffMap boolean constant, to control whether backOffMap data member,
which is created as a way to precompute probabilites, is nullified before saving the Model
instance to file (or otherwise serialized).  This can be useful for debugging or analysis.

* Added createHistBackOffMap boolean constant, which, if true, indicates to create a
set of maps from histories at level i to level i+1 (akin to backOffMap), assigning these maps
to the new histBackOffMap array data member, which will mean the maps will be saved with
the Model object when it is serialized. Such maps can be useful for debugging or analysis
(see AnalyzeDisns class).

* Added numModels method, to provide consistent interface between this and JointModel
subclass.